Removed all server classes to package server.
