+++ csc/2016/lau/lau_1/src/java/Main.java
@@ -1,3 +1,5 @@
+import server.*;
+
 public class Main {
     public static void main(String[] args) {
         Server server = new Server();
